+++ src/main/java/dev/hytalemodding/AHytaleMod.java
@@ -1,33 +1,33 @@
 package dev.hytalemodding;
 
+import com.hypixel.hytale.assetstore.map.DefaultAssetMap;
+import com.hypixel.hytale.server.core.asset.HytaleAssetStore;
 import com.hypixel.hytale.server.core.event.events.player.PlayerReadyEvent;
 import com.hypixel.hytale.server.core.plugin.JavaPlugin;
 import com.hypixel.hytale.server.core.plugin.JavaPluginInit;
 import com.hypixel.hytale.server.core.universe.world.events.AddWorldEvent;
 import com.hypixel.hytale.server.core.universe.world.events.RemoveWorldEvent;
-import com.hypixel.hytale.server.core.util.Config;
 import dev.hytalemodding.commands.ExampleCommand;
 import dev.hytalemodding.events.ChaoticMessageEvent;
 import dev.hytalemodding.events.ExampleEvent;
-import dev.hytalemodding.messages.ChaoticMessageConfig;
+import dev.hytalemodding.messages.ChaoticMessageAsset;
 import dev.hytalemodding.messages.ChaoticMessageService;
 
 import javax.annotation.Nonnull;
 
 public class AHytaleMod extends JavaPlugin {
 
-    private final Config<ChaoticMessageConfig> chaoticMessageConfig;
     private ChaoticMessageService chaoticMessageService;
+    private boolean chaoticMessageAssetsRegistered;
 
     public AHytaleMod(@Nonnull JavaPluginInit init) {
         super(init);
-        this.chaoticMessageConfig = this.withConfig("chaotic-messages", ChaoticMessageConfig.CODEC);
     }
 
     @Override
     protected void setup() {
-        this.chaoticMessageConfig.save();
-        this.chaoticMessageService = new ChaoticMessageService(this.chaoticMessageConfig);
+        registerChaoticMessageAssets();
+        this.chaoticMessageService = new ChaoticMessageService();
 
         this.getCommandRegistry().registerCommand(new ExampleCommand("example", "An example command"));
         this.getEventRegistry().registerGlobal(PlayerReadyEvent.class, ExampleEvent::onPlayerReady);
@@ -36,6 +36,20 @@ public class AHytaleMod extends JavaPlugin {
         this.getEventRegistry().registerGlobal(AddWorldEvent.class, event -> this.chaoticMessageService.trackWorld(event.getWorld()));
         this.getEventRegistry().registerGlobal(RemoveWorldEvent.class, event -> this.chaoticMessageService.untrackWorld(event.getWorld()));
         this.chaoticMessageService.startPeriodicMessages(this.getTaskRegistry());
+    }
+
+    private void registerChaoticMessageAssets() {
+        if (chaoticMessageAssetsRegistered) {
+            return;
+        }
+        getAssetRegistry().register(
+                HytaleAssetStore.builder(ChaoticMessageAsset.class, new DefaultAssetMap<>())
+                        .setPath("ChaoticMessages")
+                        .setCodec(ChaoticMessageAsset.CODEC)
+                        .setKeyFunction(ChaoticMessageAsset::getId)
+                        .build()
+        );
+        chaoticMessageAssetsRegistered = true;
     }
 
     @Override